fix: do not download an object at the exact model path
Do nothing if there is an object whose key exactly matches with the
model path. We don't need to copy that one.  For example, if the model
path is 'google/gemma-2b', we would like to copy object under
'google/gemme-2b', but not the object at 'google/gemma-2b'.